requirements.txt
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect==1.1.0
decorator==5.1.1
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1
drf-yasg==1.20.0
entrypoints==0.4
environ==1.0
executing==0.8.3
filelock==3.6.0
gensim==4.1.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
ipykernel==4.8.2
ipython==8.1.1
itemadapter==0.4.0
itemloaders==1.0.4
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
JPype1==1.3.0
jupyter-client==7.1.2
jupyter-core==4.9.2
konlpy==0.6.0
lxml==4.8.0
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.3
packaging==21.3
pandas==1.4.1
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
Protego==0.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.11.2
pymongo==4.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-scrapyd-api==2.1.2
pytz==2021.3
pyzmq==17.0.0
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy==1.8.0
Scrapy==2.6.1
scrapyd==1.3.0
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
sqlparse==0.4.2
stack-data==0.2.0
tldextract==3.2.0
tornado==6.1
traitlets==5.1.1
Twisted==22.2.0
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
w3lib==1.22.0
wcwidth==0.2.5
zope.interface==5.4.0
